Add API for engine configuration
  * Support for AAP-10647 - Show timeouts and limits in UI

+++ server/engine/datatypes.go
@@ -15,6 +15,6 @@ type Engine struct {
 	mainOutputs          *model.Output
 	done                 chan struct{}
 	status               *model.Status
-	maxExecutionRestarts int
+	maxExecutionRestarts int64
 	deploymentsClient    *armresources.DeploymentsClient
 }
